add improve of announcement
